+++ examples/backend_dart_frog/lib/jaspr_options.dart
@@ -2,8 +2,8 @@
 // Generated with jaspr_builder
 
 import 'package:jaspr/jaspr.dart';
-import 'components/counter.dart' as prefix0;
-import 'components/hello.dart' as prefix1;
+import 'package:backend_dart_frog/components/counter.dart' as prefix0;
+import 'package:backend_dart_frog/components/hello.dart' as prefix1;
 
 /// Default [JasprOptions] for use with your jaspr project.
 ///