+++ packages/jaspr_builder/lib/src/sync/sync_mixins_builder.dart
@@ -4,7 +4,7 @@ import 'package:analyzer/dart/element/element.dart';
 import 'package:build/build.dart';
 import 'package:dart_style/dart_style.dart';
 
-import '../codec/codec_resource.dart';
+import '../codec/codecs.dart';
 import '../utils.dart';
 
 /// Builds mixins for components annotated with @sync
@@ -83,59 +83,48 @@ class SyncMixinsBuilder implements Builder {
       return;
     }
 
-    var resource = await buildStep.fetchResource(codecResource);
-    var codecs = await resource.readCodecs(buildStep);
+    var codecs = await buildStep.loadCodecs();
+    var mixins = annotated.map((e) => generateMixinFromEntry(e, codecs, buildStep.inputId.toImportUrl())).join('\n\n');
 
-    var outputId = buildStep.inputId.changeExtension('.sync.dart');
-
-    var moduleImport = 'package:${buildStep.inputId.package}/${buildStep.inputId.path.replaceFirst('lib/', '')}';
-    var mixinImports = <String>{};
-
-    var mixins = annotated.map((e) {
-      var (mixin, imports) = generateMixinFromEntry(e, codecs);
-      mixinImports.addAll(imports);
-      return mixin;
-    }).join('\n\n');
-
-    var syncPart = DartFormatter(pageWidth: 120).format('''
+    var source = '''
       $generationHeader
       
       import 'package:jaspr/jaspr.dart';
-      import '$moduleImport';
-      ${mixinImports.map((p) => "import '$p';").join('\n  ')}
+      [[/]]
             
       $mixins
-    ''');
+    ''';
+    source = ImportsWriter().resolve(source);
+    source = DartFormatter(pageWidth: 120).format(source);
 
-    await buildStep.writeAsString(outputId, syncPart);
+    var outputId = buildStep.inputId.changeExtension('.sync.dart');
+    await buildStep.writeAsString(outputId, source);
   }
 
-  (String, Set<String>) generateMixinFromEntry((ClassElement, Iterable<FieldElement>) element, Codecs codecs) {
+  String generateMixinFromEntry((ClassElement, Iterable<FieldElement>) element, Codecs codecs, String baseImport) {
     final (clazz, fields) = element;
     final comp = clazz.supertype!.typeArguments.first.element!;
 
-    final imports = <String>{};
-
-    final members = fields.map((f) => """
-      ${f.type.getDisplayString()} get ${f.name};
-      set ${f.name}(${f.type.getDisplayString()} ${f.name});
-    """).join('\n');
+    final members = fields.map((f) {
+      var type = codecs.getPrefixedType(f.type);
+      return """
+        $type get ${f.name};
+        set ${f.name}($type ${f.name});
+      """;
+    }).join('\n');
 
     final decoders = fields.map((f) {
       var decoder = codecs.getDecoderFor(f.type, "value['${f.name}']");
-      imports.addAll(decoder.$2);
-      return "${f.name} = ${decoder.$1};";
+      return "${f.name} = $decoder;";
     }).join('\n');
 
     final encoders = fields.map((f) {
       var encoder = codecs.getEncoderFor(f.type, f.name);
-      imports.addAll(encoder.$2);
-      return "'${f.name}': ${encoder.$1},";
+      return "'${f.name}': $encoder,";
     }).join('\n');
 
-    return (
-      """
-      mixin ${clazz.name.startsWith('_') ? clazz.name.substring(1) : clazz.name}SyncMixin on State<${comp.name}> implements SyncStateMixin<${comp.name}, Map<String, dynamic>> {
+    return """
+      mixin ${clazz.name.startsWith('_') ? clazz.name.substring(1) : clazz.name}SyncMixin on State<[[$baseImport]].${comp.name}> implements SyncStateMixin<[[$baseImport]].${comp.name}, Map<String, dynamic>> {
         $members
       
         @override
@@ -156,8 +145,6 @@ class SyncMixinsBuilder implements Builder {
           SyncStateMixin.initSyncState(this);
         }
       }
-    """,
-      imports
-    );
+    """;
   }
 }