+++ packages/jaspr_builder/test/codec/sources/model_class.dart
@@ -1,11 +1,11 @@
 import 'dart:convert';
 
 const modelClassSources = {
-  'site|lib/model.dart': '''
+  'site|lib/model_class.dart': '''
     import 'package:jaspr/jaspr.dart';
         
-    class Model {
-      Model(this.a, {this.b = 18, this.c, required this.d});
+    class ModelA {
+      ModelA(this.a, {this.b = 18, this.c, required this.d});
       
       final String a;
       final int b;
@@ -13,8 +13,8 @@ const modelClassSources = {
       final bool d;
       
       @decoder
-      factory Model.fromRaw(Map<String, dynamic> raw) {
-        return Model(model['a'], b: model['b'], c: model['c'], d: model['d']);
+      factory ModelA.fromRaw(Map<String, dynamic> model) {
+        return ModelA(model['a'], b: model['b'], c: model['c'], d: model['d']);
       }
       
       @encoder
@@ -26,15 +26,13 @@ const modelClassSources = {
 };
 
 final modelClassOutputs = {
-  'site|lib/model.codec.json': jsonEncode({
-    "id": ["site", "lib/model.dart"],
+  'site|lib/model_class.codec.json': jsonEncode({
     "elements": [
       {
-        "name": "Model",
-        "extension": null,
+        "name": "ModelA",
         "decoder": "fromRaw",
         "encoder": "toRaw",
-        "import": "package:site/model.dart"
+        "import": "package:site/model_class.dart",
       }
     ]
   }),