+++ packages/jaspr_builder/test/codec/sources/model_extension.dart
@@ -1,22 +1,26 @@
 import 'dart:convert';
 
 const modelExtensionSources = {
-  'site|lib/model.dart': '''
+  'site|lib/model_type.dart': '''
     import 'package:jaspr/jaspr.dart';
         
-    class Model {
-      Model(this.a, {this.b = 18, this.c, required this.d});
+    class ModelB {
+      ModelB(this.a, {this.b = 18, this.c, required this.d});
       
       final String a;
       final int b;
       final double? c;
       final bool d;
     }
-    
-    extension type ModelCodec._(Model model) implements Model {  
+  ''',
+  'site|lib/model_extension.dart': '''
+    import 'package:jaspr/jaspr.dart';
+    import 'model_type.dart';
+        
+    extension type ModelBCodec(ModelB model) implements ModelB {  
       @decoder
-      factory ModelCodec.fromRaw(Map<String, dynamic> raw) {
-        return ModelCodec._(Model(model['a'], b: model['b'], c: model['c'], d: model['d']));
+      factory ModelBCodec.fromRaw(Map<String, dynamic> raw) {
+        return ModelBCodec._(ModelB(model['a'], b: model['b'], c: model['c'], d: model['d']));
       }
       
       @encoder
@@ -28,15 +32,15 @@ const modelExtensionSources = {
 };
 
 final modelExtensionOutputs = {
-  'site|lib/model.codec.json': jsonEncode({
-    "id": ["site", "lib/model.dart"],
+  'site|lib/model_extension.codec.json': jsonEncode({
     "elements": [
       {
-        "name": "Model",
-        "extension": "ModelCodec",
+        "name": "ModelB",
+        "extension": "ModelBCodec",
         "decoder": "fromRaw",
         "encoder": "toRaw",
-        "import": "package:site/model.dart"
+        "import": "package:site/model_extension.dart",
+        "typeImport": "package:site/model_type.dart",
       }
     ]
   }),