+++ packages/jaspr_builder/test/styles/css_annotation_test.dart
@@ -1,8 +1,10 @@
 import 'package:build/build.dart';
 import 'package:build_test/build_test.dart';
+import 'package:jaspr_builder/src/styles/styles_bundle_builder.dart';
 import 'package:jaspr_builder/src/styles/styles_module_builder.dart';
 import 'package:test/test.dart';
 
+import 'sources/bundle.dart';
 import 'sources/styles_class.dart';
 import 'sources/styles_global.dart';
 
@@ -28,6 +30,18 @@ void main() {
           reader: await PackageAssetReader.currentIsolate(),
         );
       });
+    });
+
+    test('generates bundle', () async {
+      await testBuilder(
+        StylesBundleBuilder(BuilderOptions({})),
+        {
+          ...stylesClassOutputs,
+          ...stylesGlobalOutputs,
+        },
+        outputs: stylesBundleOutputs,
+        reader: await PackageAssetReader.currentIsolate(),
+      );
     });
   });
 }